v1.2.0

+++ src/folders/folders.controller.ts
@@ -7,16 +7,24 @@ import {
   Post,
   Put,
   Res,
+  UseGuards,
 } from '@nestjs/common'
+import { plainToInstance } from 'class-transformer'
 import { Response } from 'express'
 import { AppService } from 'src/app.service'
-import { CreateFolderDto, UpdateFolderDto } from './dto'
+import {
+  FolderAlreadyPresentException,
+  FolderNotFoundException,
+} from 'src/exception'
+import { JwtAuthGuard } from 'src/users/jwt/jwt-auth.guard'
+import { CreateFolderDto, FolderDto, UpdateFolderDto } from './dto'
 import { FoldersService } from './folders.service'
 
 @Controller({
   path: 'folders',
   version: '1',
 })
+@UseGuards(JwtAuthGuard)
 export class FoldersController {
   constructor(
     private readonly foldersService: FoldersService,
@@ -25,18 +33,31 @@ export class FoldersController {
 
   @Post()
   async create(@Res() response: Response, @Body() body: CreateFolderDto) {
+    const folderCheck = await this.foldersService.findFolderByName(body.name)
+    if (folderCheck) throw new FolderAlreadyPresentException()
+
     const folder = await this.foldersService.create(body)
     return this.appService.okFolderCreated(response, folder.uuid)
   }
 
   @Get()
   async findAll() {
-    return this.foldersService.findAllFolders()
+    const folders = await this.foldersService.findAllFolders()
+
+    const mappedFolders = plainToInstance(FolderDto, folders)
+
+    return mappedFolders
   }
 
   @Get(':id')
   async findFolder(@Param('id') id: string) {
-    return this.foldersService.findFolder(id)
+    const folder = await this.foldersService.findFolder(id)
+
+    if (!folder) throw new FolderNotFoundException()
+
+    const mappedFolder = plainToInstance(FolderDto, folder)
+
+    return mappedFolder
   }
 
   @Put(':id')
@@ -45,12 +66,24 @@ export class FoldersController {
     @Param('id') id: string,
     @Body() body: UpdateFolderDto
   ) {
+    const folderExists = await this.foldersService.findFolder(id)
+    if (!folderExists) throw new FolderNotFoundException()
+
+    const folderNameCheck = await this.foldersService.findFolderByName(
+      body.name
+    )
+    if (folderNameCheck) throw new FolderAlreadyPresentException()
+
     await this.foldersService.update(id, body)
+
     return this.appService.okFolderUpdated(response, id)
   }
 
   @Delete(':id')
   async delete(@Res() response: Response, @Param('id') id: string) {
+    const folderCheck = await this.foldersService.findFolder(id)
+    if (!folderCheck) throw new FolderNotFoundException()
+
     await this.foldersService.delete(id)
     return this.appService.okFolderDeleted(response, id)
   }
